renames package to depend
