+++ packages/depend/.babelrc
