+++ packages/depend/.eslintignore
