+++ packages/depend/.eslintrc.json
