+++ packages/depend/.npmignore
