+++ packages/depend/__tests__/__mocks__/fs.js
