+++ packages/depend/__tests__/cli.test.js
