+++ packages/depend/__tests__/fixtures/divergent.json
