+++ packages/depend/cli.js
