+++ packages/depend/depend.js
