+++ packages/depend/pat-regexp.js
