+++ packages/depend/strategies.js
